@@ -9,7 +9,7 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.1",
         "guzzlehttp/psr7": "^2.0"
     },
     "autoload": {
@@ -20,5 +20,9 @@
     "require-dev": {
         "squizlabs/php_codesniffer": "^3.6",
         "vimeo/psalm": "^4.20"
+    },
+    "scripts": {
+        "cs": "phpcs --standard=PSR12 ./src/",
+        "analyze": "psalm --show-info=true"
     }
 }
